resotocore/pyproject.toml
[tool.poetry]
name = "resotocore"
version = "2.0.0a17"
description = "Resoto core."
authors = ["Some Engineering Inc."]
license = "Apache-2.0"
homepage = "https://resoto.com/"
repository = "https://github.com/someengineering/resoto"
documentation = "https://resoto.com/docs/"
keywords = ["cloud", "security", "infrastructure intelligence"]
classifiers = [
    # Current project status
    "Development Status :: 4 - Beta",
    # Audience
    "Intended Audience :: System Administrators",
    "Intended Audience :: Information Technology",
    # License information
    "License :: OSI Approved :: Apache Software License",
    # Supported python versions
    "Programming Language :: Python :: 3.8",
    # Supported OS's
    "Operating System :: POSIX :: Linux",
    "Operating System :: Unix",
    # Extra metadata
    "Environment :: Console",
    "Natural Language :: English",
    "Topic :: Security",
    "Topic :: Utilities",
]
include = ["resotocore/static/*"]

[tool.poetry.scripts]
resotoworker = 'resotoworker.__main__:main'

[tool.poetry.dependencies]
python = "^3.8"
aiohttp-swagger3 = "0.7.0"
aiohttp = { extras = ["speedups"], version = "3.8.1" }
jsons = "1.6.1"
parsy = "1.4.0"
plantuml = "0.3.0"
python-arango = "7.3.0"
python-dateutil = "2.8.2"
toolz = "0.11.2"
transitions = "0.8.10"
APScheduler = "3.8.1"
aiostream = "0.4.4"
tzlocal = "4.1"
frozendict = "2.1.3"
PyYAML = "6.0"
jq = "1.2.1"
posthog = "1.4.5"
# available in python 3.8
ustache = "0.1.5"
aiofiles = "0.8.0"
cryptography = "36.0.1"
rich = "11.1.0"
resotolib = { path = "../resotolib", develop = true }

[tool.poetry.dev-dependencies]
flake8 = "4.0.1"
shared_dev = { path = "../shared_dev" }
pytest-asyncio = "0.17.2"
deepdiff = "5.7.0"
hypothesis = "6.36.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.paths]
source = ["resotocore"]

[tool.coverage.run]
branch = true
concurrency = ["multiprocessing"]
source = ["resotocore"]
omit = ["*/resotolib/*"]

[tool.coverage.report]
show_missing = true
